Enhance: Use flux for state management
Add flux package

Add `TodoStore` where we put all the actions relating to the storing of and state of Todo items

Add `TodoActions` which calls to dispatcher to run actions from the `TodoStore`. This is what any functions that need an operation done on Todo call.

Add `TodoConstants` which stores the keys for available actions

Add `ETDispatcher` which is a custom dispatcher

Add ulid package for "better" IDs for items

Fix bug where `onComplete` method is not passed to  the `TodoItem` component

Fix controlled input bug on complete checkbox in `TodoItem` component by using `defaultCheck` instead of `checked` attribute