Feat: Add ability to delete and edit todos
Add delete and update methods to `App`. These methods call `TodoActions` to dispatch the relevant event. The method are passed down to the `TodoItem` component

Add a functional component `Editable`, which allows is to edit a `Todo` inline.

Add a bin icon (svg image) in public/images to use for the delete icon

Add delete and update functions and event handler code to `TodoStore`

Note: When editing, a todo will only prompt to save if the content changes